% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust2mat.R
\name{gforce.clust2mat}
\alias{gforce.clust2mat}
\title{Convert a clustering or grouping to partnership matrix.}
\usage{
gforce.clust2mat(clusters)
}
\arguments{
\item{clusters}{length \eqn{d} vector. Assigns each variable or data point to a cluster. Cluster names
can be numbers or strings.}
}
\value{
a \eqn{d x d} numeric array that contains the partnership matrix corresponding to \code{clusters}.
}
\description{
Takes a clustering \eqn{G} and constructs \eqn{B(G)}.
}
\examples{
clusters <- c(1,1,1,2,2,2,3,3)
B_clust <- gforce.clust2mat(clusters)
}
