% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDR_control.R
\name{gforce.confint2test}
\alias{gforce.confint2test}
\title{Convert confidence intervals to equivalent test statistics.}
\usage{
gforce.confint2test(conf_ints, alpha)
}
\arguments{
\item{conf_ints}{\eqn{d x d x 3} array. Each \eqn{d x d x 1} slice is a symmetric matrix.}

\item{alpha}{confidence level level of the confidence intervals.}
}
\value{
a \eqn{d x d} symmetric matrix of test statistics.
}
\description{
Can convert a 4D array encoding the confidence intervals for a precision
matrix to standard normal test-statistics.
}
