% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_SDP_matrix.R
\name{gforce.kmeans_SDP_matrix}
\alias{gforce.kmeans_SDP_matrix}
\title{K-means SDP Matrices.}
\usage{
gforce.kmeans_SDP_matrix(D, K)
}
\arguments{
\item{D}{\eqn{d \times d} numeric matrix.}

\item{K}{number of clusters.}
}
\description{
Constructs the SDP constraint matrices so they can be passed to a black-box solver.
}
